requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
numpy
matplotlib
tqdm